+++ src/main/java/cn/heyuantao/devicemanagement/repository/TypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * @author he_yu
  */
-public interface TypeRepository extends JpaRepository<Type,Integer> {
+public interface TypeRepository extends JpaRepository<Type,Long> {
     /**
      *
      * @param name
